Add bytepipe code + tests
Bytepipe mode works in two ways so far:

- As a process
- As a function you can enter

The behaviour of each is the same, except on termination, where specific
callbacks can be set to change how to react to closed or erroring out
sockets. Because of this, the process form does not adhere to OTP
standards outside of supporting proc_lib calls for synchronous starts.

The module also expects a ranch-like use of behaviours, and it could
theoretically be possible to pipe TCP ports into UDP ones, and
vice-versa.

Tests are non-exhaustive at this point, but sufficient to demonstrate
that the base mechanism works.