Support chunked streaming
This is similar to regular parsing, but on partial results (where the
input binary is not enough to finish parsing the current chunk) the
function will instead return the buffer it went through and a
continuation for its next state.

The continuation can then be used with further binary data and the
parsed chunks can be flushed.

This allows to know what the delimitations of a chunked-encoded message
is without needing to buffer it entirely in memory before forwarding it
to a peer.