Consistency in order of arguments
The usage of Req and State is all over the place and very confusing,
somewhat error prone. This is mostly my fault, too.

This patch makes it so that Req and State are always the last arguments
(except in cowboy functions that take an optional default value), and
Req always comes before State.