Handling chunked to HTTP/1.0 stream decoding
Chunked encoding isn't supported by 1.0 clients, and should be handled
to be compliant.

This patch adds a copy of 'vegur_chunked' that does the same thing, but
does not keep the framing information -- it actually decodes chunked
content.

The proxy logic now detects when sending chunked data to a 1.0 client
and decodes the data. However, to avoid accumulating the data
indefinitely, we make use of the HTTP feature where we can omit a
content-length, add connection:close, and use the connection termination
as a marker for the end of content.