All undef length responses are stream_close
According to newer specs like HTTP Bis, all responses without a
specified content-length or transfer-encoding should be considered
to be closing on connection close, whether they are keep-alive or not.

Given most of our own calls also feature a connection:close header using
this client, it is somewhat safe to assume the connection will be closed
no matter what, although this could be a risk to assume browsers will
behave the same.

This patch makes it so that whenever we don't have a content-size
specified, we assume a body delimited by the end of a connection. Other
bodies *should* include a content-length: 0 or a chunked with only a
final chunk.