Fix bytepipe with proxy protocol
Proxy protocol uses a non-transparent port object that makes matching
explicitly on the port a non-starter for the client connection.

This fix moves the matching of the client connection down after the dyno
connection, which is likely to be TCP and delays the protocol changes
required until this is modified.