Add optional per-endpoint 100-continue bypass
The bypass requires moving the continue middleware execution after app
lookup (otherwise, no way to do it per-app!), and the addition of a new
callback that will be usable for other specific or tweakable features on
the basis of a 'disabled' | 'enabled' return value.

The bypass added here is for 100-continue. When enabled, the continue
middleware will send the '100 continue' response on the line, remove the
header from the endpoint's view, and keep handling things fine
otherwise.