Add default per-endpoint 100-continue bypass
The bypass requires moving the continue middleware execution after app
lookup (otherwise, no way to do it per-app!), and the addition of a new
argument when looking up a domain that will be usable for other specific or
tweakable features on the basis of a 'deep_continue' option being
passed.

The bypass added here is for 100-continue. By default, the continue
middleware will send the '100 continue' response on the line, remove the
header from the endpoint's view, and keep handling things fine
otherwise.

By passing the [deep_continue,...] Option list as part of the lookup,
the application can instead favorise the end-to-end approach to
100-continue handling.