Add checkin phases
Because error reasons can be the same in multile contexts (for example,
timeout in an established connection, vs. timeout on a connection being
established), this patch adds in the concept of 'phases' to make it
simpler to understand which action to take depending on context.

Two phases are declared for now: connecting and connected.