When a given event >1 instance, return newest
Before this patch, whenever an event log would hold duplicate events
(i.e. multiple reconnections) and that a value was fetched, the logging
system pretended no event were found (it was 1 or 0).

This fix makes it so that the event logger understands that multiple
copies of an event can be found, and reports the newest one instead (the
assumption here is that older events were to be logged anyway).

This commit may need expanding if we decide we need to log based on
oldest events of a given key.