Replace custom raw funs with internal-API calls
both raw_socket and raw_sockbuff are dependencies on a custom cowboy
fork. The function calls can be cleaned up and replaced with uses of the
internal API (used for hooks and whatnot, guaranteed not to go away by
Cowboy's author in a face-to-face conversation).

This should reduce our dependencies on custom patches to the server long
term.