Allow variying receive buffer sizes
Variable buffer is required to allow endpoints to send more or less data
without depending on what the vegur router can handle.

This patch makes it configurable, and makes tests work around the idea
that limitations for header sizes (non-cookie related) are not going to
be enforced if 100% of the header is included within that buffer (we
were committing to that much space anyway).

The default size is set to 1MB.