closing an in-progress chunked conn leaves data on
The chunked mechanism in vegur makes use of a continuation wrapped along
with the #client{} record to transparently work across packet
boundaries.

However, when the connection is killed, the client along with the
continuation is passed to vegur_client:close/1, which ends up failing
with a function_clause error.

This patch makes it so that close/1 expects the continuation to be there
as part of regular operations to keep the abstraction going and silence
the errors.