Refactoring to allow cleaner status return meta
The metadata was refactored as a quickfix so that vegur_req wouldn't
return 0 for most request return values, but doing so was at the cost of
making the concerns of meta code usage external to the request log where
it was restricted, and didn't solve the case of other middleware
interrupting a request with a given status code.

This patch adds in a return path for `{halt, Code, Req}` (on top of the
existing `{halt, Req}`) used for terminating a request that allows to
return the code that will be used for metrics and statistics down the
line.