Handle failed downstream connection in raw_request
The connection may be interrupted while sending a piece of data to the
back-end and should be reported as such rather than a crash.

A connection can be interrupted and noticed by the proxy before the
client receives any data. Therefore we shouldn't wait on data making it
through the pipe.