Eliminating no_return warning
The no_return was due to invalid specifications of options expected,
which combined with subtyping rules in Dialyzer (I think?) led to
Dialyzer assuming the call was invalid and would fail or loop forever.

Adding all the options to the spec broadened the subtyping rules to work
with the usage of vegur_bytepipe:become that vegur_proxy does and
cleared up a warning.