Handle invalid HTTP headers
Headers can have invalid values according to the HTTP spec, such as a
request sending in 'expect: 100 continue' headers (the canonical value
requires a dash (-)).

This patch makes it so that the header parsing functions expect such bad
header values and respond correctly with a 400 error when they happen.

A test was also added to verify the 417 behavior of bad expects on
continue requests.