Detect upstream connections closed on responses
When sending data in passive mode, it is usually impossible to be
notified of connections being closed. For this to be done, we need to
poll the socket we're writing to.

The patch adds this functionality, at the cost of breaking pipelining
(which isn't too used anyway, and was never documented as supported).

The handling of idle data on the wire is to be revisited -- right now we
interrupt everything and close with a 400.