Read 1 byte per sent packet, do less work
The buffer is switched to a binary to avoid having to dynamically go
through a list all the time (we may incur more transforms on a binary
however, but cowboy does it everywhere), and we are reading one byte
from the client per packet sent by the server to reduce the chances of
filling a buffer before the response is done by a honest client.

Also only count the size of the buffer once we know we'll have to modify
it.