Do not respect close-delimitation on HEAD,204,304
It's common practice for people to omit content-lengths on
identity-encoded responses for HEAD requests, and status codes that do
not accept bodies such as 204 and 304. In such cases, respect the
requester's connection setting.

For example, a HEAD request from a keep-alive request will remain
keep-alive even if it uses an identity transfer-encoding (Default for
HTTP), and same for a 204 or 304 response.

However, if the request had a 'connection: close' header, the response
shall be forced to close anyway, nevermind what the endpoint asked for,
given the client told us the intent not to have multiple
requests/responses over the line.