Handle packet-bound trailing \r\n in chunked http
This handles a specific corner-case of chunked pieces of content that
are straddling packed boundaries in all cases (as before) more
particularly the case of 0-lenght chunks without the appropriate number
of CLRFs (a single one instead of two) that we have decided to handle to
respect what happens in the wild for many implementations.

This code makes it so that even though we're parsing byte-by-byte,
recognizing a lone 0\r\n from a 0\r\n\r\n even if the second \r\n
happens to be on a packet boundary and requires further reading from a
socket, we should get the right amount of CLRFs to respect the original
requests.

There are caveats regarding how long we're willing to wait -- in cases
where the follow-up packet isn't in the TCP socket's recv buffer
already, we give up on waiting early and will send in a lone \r\n, which
most browsers understand anyway.