Add event-log merging capability to vegur_req_log
This will allow to avoid having to string a single event log through all
modules, and instead open up the door to the idea of running one event
log per module, then later merging them together.

The risk of this practice, given we use os:timestamp(), is that we
introduce non-linearity in the logs. It is for example possible that two
events get the same time stamp when they're sequential and the merge
swaps them in order.

This is partly alleviated through a merging algorithm that tries to
order similar items consistently, but does not guarantee the order in
execution and the order in logs in any absolute sense.

This is likely a nicer functionality when using the req_log for a
key-based search rather than through linear reports.