Add data-returning capability to streaming funs
In order to gather statistics from vegur_client in streaming functions
dedicated to cowboy, we need to either:

- change cowboy to allow returns
- use the process dictionary (as we do for TCP buffers)
- use a throw and escape the tyranny of cowboy's handling

This allows to throw the value out of it and fetch the client back, and
opens the door to moving away from process dictionaries for buffers.

The downside of it is that this forever disables the use of OnResponse
callbacks as part of vegur if we expect them to be having an impact in
any way on `Req'.