Go back to a non-throw based communication
Despite the fact that the comments mention needing to throw to get
our data back and out of the delivery mechanism, it appears removing it
has no impact on our ability to fetch logged data from there.

I'm still not sure how it manages to work without it, but it apparently
does.

Plus we don't need to break cowboy apparently?