Remove trailing errors on interrupting errors
This change introduces the function `borrow_cowboy_socket/1`, a
varition of `raw_cowboy_socket/1`, which has been fixed for edge
behaviours.

The `raw_cowboy_socket/1` function was intended for the proxy taking
over cowboy when it came to requests. For this reason, it tried to mark
the request as closed, but had forgotten to send the internal cowboy
message used to deal through abnormal exception flow -- this is fixed
both there and in `raw_cowboy_sockbuf/1`.

In the case of `100 Continue` responses, the intended use was to add
data to the connection, but without actually taking over the connection
permanently. Marking it as closed could cause issues, and for this
reason, the value was ignored.

However, when fixing the behaviour of `raw_cowboy_socket/1`, the marking
of the connection became impure, yielding early termination in such
responses. The `borrow_cowboy_socket/1` function is added to represent
the original intended use from the existing one.

Tests have been altered to reflect this change.