Fix large uploads interrupted by eager response
When a response comes faster than a large body is being uploaded, with the
backend server then closing its connection rapidly (as is the case with
say, a 403 response to a file upload), vegur may detect the connection
termination as an error and ignore the response on the wire.

This fix adds a 64kb buffer for an early response that will allow vegur
to recognize and accumulate an early response.

In ranch transports where data on the line remeains readable for as long
as possible, this will allow Vegur to push the response as usual after
figuring out it was all a big misunderstanding.

In ranch transports where the connection being closed may shut down the
ability to read the data queued up, this fix will allow to tolerate up
to 64kb of data without a problem.

Given interrupting responses are often error codes (401, 403, 404),
redirections (3xx), or server errors (5xx), and that these tend to have
a limited content, this should cover most cases without much of a
problem.

The buffer size can be made bigger if this is still an issue.

Note that this fix does *not* speed up the time it may take for the
underlying implementation to figure out a connection has been closed --
this may still take a while. It just makes it so the least amount of
data possible gets lost in the communication.