Be strict on the termination of chuncks
Chuncks currently can terminate in a `0\r\n` on a packet boundary.
When that happens, we poll for data on the pipe once, and then give up,
sending the last terminal chunk that way.

This however causes problems on strict servers (and clients) that end up
waiting for data that never comes down the line.

This pull request changes things so that Vegur will wait on a `0\r\n`
and keep polling for a final `\r\n` without giving up until it gets it
(or garbage on the line).

Specific changes in behaviour:
- a new error value for `bad_chunk` has been added in the form of
  `{bad_chunk, {last_crlf, _Buffer}}` and will need to be handled by
  libraries using Vegur
- A request without the `\r\n` missing at the end will now wait for the
  default amount of time before a timeout (defaults at 55s)
- The server will receive partially sent chunks that may still hang
  the way they were doing it before, but will return a different
  error code, blaming the client instead of the server
- If garbage follows the bad chunk (trailers without the header, next
  request), the request will explicitly fail with a bad_chunk error;
  the server may or may not have the time to see content coming back
  up, and should never have the time to send a response.
- In the case above where the server sends an early response, the
  current state of thing is that this response is buffered and won't
  show up.
- trailers are still only supported if the trailers header is there.
  In the case the header wasn't included, we change from forwarding an
  incomplete chunk stream to erroring out.