Expose the response headers through vegur_req
Since we enabled to set headers on responses from the dyno to the
client, I figure it would also be useful to enable the headers set by
the client to be visible to the routing layer.

The headers are unmodified, and rather than replicating the existing
`cookie(Key, Req)` mechanism, this one returns an unmodified resource.

This is specifically because the server can set some headers multiple
time, with cases such as `Set-Cookie: ...` being there once per header
(this issue doesn't exist client-side because the `Cookie` header is
meant to merge all values in one header).