Adding a breaking test for HEAD conn handling
The connection should be left open in order for HEAD to be optimal in
usage -- the connection, despite having no length, doesn't always have a
streaming body when it's retried without HEAD. THis is often just an
optimization made by web servers we proxy for, which avoid to do the
full processing for a request and return undefined lengths.

In such circumstances, rather than faithfully reproducing the
'Connection: close' that we would have created in such circumstances, we
have the option of forcing the connection to remain open.

We already had the headers set properly in the response, but the
handling of the connection is interrupted by cowboy.

A pull request exists to help this commit pass its test, following an
update to rebar.config:

https://github.com/heroku/cowboy/pull/11