Shorter poll loop
This reduces the polling interval of content bodies from client to
55 1 second attempts interspersed with backend checks, rather than a
single sequence of 55 seconds.

Under the current implementation, the connection to the back-end server
is polled after every packet to detect connection terminations. This
works fine to rapidly find interruptions as long as the front-end client
is sending packets.

However, in the off-chance a client (mobile or malevolent) goes silent
for extended periods of time (up to 55s), vegur is blocked waiting on
the recv call on that client socket, and cannot poll or detect
interruptions from the back-end server as rapidly as it could on a
loopback interface, for example. This can lead to closed connections
artificially hanging for up to 55 seconds in the proxy, and can be used
as a kind of weird DoS mechanism.

This patch makes it so the maximal period of time of inactivity between
polls is now 1 second; still long, but far less dramatic than it was.

Because it is not possible to properly test this change on a loopback
interface, this patch includes a change to our test structure that tries
to use a non-loopback interface when available. It will possibly make no
difference on localhost, but some integration servers (TracisCI) will
somehow be able to be more representative of a real world scenario.

The dependency on a new cowboy branch is pending a merge and tagging in
our repository.