Fix bad poll loop logic
The previous iteration had a few important problems:

- Some requests would fail after 1s when the first packets only were late.
- Waiting longer on the first packet without polling the other end of
  the connection delayed detecting connection terminations
- 1s poll timeouts would end up repeating buffered data

This patch reworks the flow of session streaming to always account for
all packets equally, always giving the 55s timeout with 1s poll
intervals.

This is done by pretending to send data when we have an empty buffer --
so it is not sent, as before -- and otherwise doing the same routine as
non-first packets.

The end-result is identical behaviour with regards to how we send first
packets (only when we have some to send), but corrected behaviour on 1s
polling sequences (no short failure on first packet) while adding
earlier detection of failures (we can detect them *before* the first
body packet is received) and avoiding repeated packets being uploaded.