Rework the phases in checkins
Currently the `checkin_service/6` callback is called in the following
places:
- in `vegur_lookup_service_middleware` when a checkout succeeds but a
  connection fails (Phase: `connecting`, with a reason given for
`ServiceState`);
- in `vegur_midjan_middleware` when a connection is over (Phase:
  `connecting`, with a hardcoded `normal` as a Service State)

The problem is that whenever an error happened at the `checkout_service`
part of the interface rather than on the proxying level, errors have to
be handled in the `checkin_service` callback under the `connected` phase
(which never happened).

This refactors and expands the interface to either be clearer by
supporting more descriptive phases (`connecting`, `connected`,
`checkout`, `lookup`) and by carrying error reasons as the service state
as returned by the interface module.

This represents a breaking change.