Revert Transport:recv(_,_,0) -> 1
Having a 1 in the timeout could result in consequences in the
scheduler. they're unavoidable as a workaround for the ssl module, but
if it's just a tcp socket, it shouldn't have to suffer.

The fix for catching this is in release 1.3.2 of the
ranch_proxy_protocol library, which this commit locks to.