Stop relying on ENV var for keepalive
The keepalive value is now handled by having a different possible return
value from the old one. `{Ip, Port}` as a service value keeps its
original meaning while `{keepalive, {default, {Ip, Port}}}` has been
added to mean "keepalive to the previous port; if not possible use the
`{Ip, Port}` combo that follows".

Such a value is then kept as possible to keep-alive to for subsequent
calls. A call that isn't marked as keepalive will never be keepalive
even if a subsequent request asks for it.