Allow proper replacement of keepalive connections
In cases where the caller wants to do keepalive to the back-end with
limits on how many connections may be used (or may want to force a
replacement node to take the follow-up requests), the former semantics
only allowed to declare a continuing keepalive connection or a
non-keepalive one.

This yielded a weird case where if we have, say, 15 requests over a
client connection and want to send batches of 10 over keep-alive to a
back-end, the pattern required to do:

keepalive: 1..10
non-keepalive: 11
keepalive : 12..15

So instead of a more optimal path, the current API mandated the use of
an extra connection every time the limit was attained to avoid forcing a
keepalive conn.

This commit includes the option of specifying keepalive as either:

- {keepalive, {default, Backend}}
- {keepalive, {new, Backend}}

The usage of `new` here mandates forced closure of the backend currently
connected to, and calls for opening a new connection in its place that
will be ready for keepalive.

This allows proper support for a sequence being:

keepalive: 1..10
keepalive: 11..15

With only two connections.

A test has been added for the handling there.