Enable SO_REUSEADDR for backend connections
The Erlang TCP stack translates a TCP `connect` call to a set of
standard operations in the linux kernel:

- open a socket
- bind(Socket, "0.0.0.0", 0)
- connect(Socket, Ip, Port)

Surprisingly enough, that `bind` call always happens and has the
unfortunate side-effect of picking one of the currently unassigned ports
in the ephemeral port range.

The negative aspect here is that despite TCP identifying connections by
4-tuple (`{LocalIp, LocalPort, RemoteIp, RemotePort}`), which would let
us use the entire ephemeral port range for each remote `IP:Port` combo,
the way Erlang uses them forces us to only the ephemeral port range due
to the `bind` call's limitations.

By passing in `SO_REUSEADDR` (or in the case of Erlang, `{reuseaddr,
true}`) we allow the bind call to reuse the ephemeral port. The check
will then be done by the connect syscall to see if the value is valid or
not (against the 4-tuple mapping). If the value is not valid, the error
returned will then be the POSIX code for `EADDRNOTAVAIL` (or just
`eaddrnotavail` in Erlang).

This frees up the full port range and isolates one endpoint being
overloaded from others suffering the same fate.