Remove X-Forwarded-Host headers
Some backend applications vegur is used to proxy through may give a
higher priority to `X-Forwarded-Host` than `Host` as a header, and this
can be used in part of attempts of setting up cache poisoning.

This patch makes sure that the header is unset and not passed to backend
applications.