Handle temporary buffer right on backend close
The calculation to know if we needed to forward data from the back-end
to the front-end only used the buffer accumulated prior to the
connection being closed or reset. When all the response data fit in one
packet received right before the termination, the response would be
invisible and ignored.

This patch considers both the prior and the most recent buffer to be
non-empty, ensuring that on the short span, the response is actually
forwarded. NewDownBuffer contains DownBuffer, so checking only the
former contains the latter.

This has been a long-standing very subtle bug!

Also contains minor fixes to the demo router.